requirements.txt
flask==2.3.3
azure-storage-blob==12.17.0
azure-ai-formrecognizer==3.2.1
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
python-dotenv==1.0.0 